+++ apps/backoffice-v2/src/common/components/atoms/XCircle/XCircle.tsx
@@ -10,7 +10,7 @@ export interface IXCircle extends Omit<LucideProps, 'size'> {
 export const XCircle: FunctionComponent<IXCircle> = ({ containerProps, size = 24, ...props }) => {
   return (
     <IconContainer {...containerProps} size={size}>
-      <X {...props} size={size * 0.55} className={ctw('stroke-[4px]', props.className)} />
+      <X {...props} size={size * 0.55} className={ctw('stroke-1', props.className)} />
     </IconContainer>
   );
 };