+++ apps/backoffice-v2/src/domains/workflow-definitions/fetchers.ts
@@ -29,6 +29,7 @@ export const WorkflowDefinitionConfigSchema = z
     isAssociatedCompanyKybEnabled: z.boolean().default(false),
     isCaseOverviewEnabled: z.boolean().default(false),
     isCollectionFlowPageRevisionEnabled: z.boolean().default(false),
+    isAgentEditingEnabled: z.boolean().default(false),
     isCaseRiskOverviewEnabled: z.boolean().default(false),
     isDocumentTrackerEnabled: z.boolean().default(false),
     theme: WorkflowDefinitionConfigThemeSchema.default({