+++ apps/backoffice-v2/src/domains/workflows/hooks/mutations/useUpdateWorkflowByIdMutation/useUpdateWorkflowByIdMutation.tsx
@@ -19,7 +19,8 @@ export const useUpdateWorkflowByIdMutation = ({ workflowId }: { workflowId: stri
         | 'ask_revision_document'
         | 'update_document_properties'
         | 'step_request'
-        | 'step_cancel';
+        | 'step_cancel'
+        | 'edit_collection_flow';
     }) =>
       fetchUpdateWorkflowById({
         workflowId,