+++ apps/backoffice-v2/CHANGELOG.md
@@ -4,14 +4,24 @@
 
 ### Patch Changes
 
-- @ballerine/workflow-browser-sdk@0.6.111
-- @ballerine/workflow-node-sdk@0.6.111
+- Bump
+- Updated dependencies
+  - @ballerine/ui@0.7.133
+  - @ballerine/workflow-browser-sdk@0.6.112
+  - @ballerine/workflow-node-sdk@0.6.112
+  - @ballerine/react-pdf-toolkit@1.2.104
+  - @ballerine/workflow-browser-sdk@0.6.111
+  - @ballerine/workflow-node-sdk@0.6.111
 
 ## 0.7.132
 
 ### Patch Changes
 
+- Bump
 - Updated dependencies
+  - @ballerine/common@0.9.89
+  - @ballerine/workflow-browser-sdk@0.6.111
+  - @ballerine/workflow-node-sdk@0.6.111
   - @ballerine/ui@0.7.132
   - @ballerine/react-pdf-toolkit@1.2.103