+++ examples/headless-example/CHANGELOG.md
@@ -1,9 +1,18 @@
 # @ballerine/headless-example
 
+## 0.3.111
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.112
+
 ## 0.3.110
 
 ### Patch Changes
 
+- Updated dependencies
+  - @ballerine/common@0.9.89
+  - @ballerine/workflow-browser-sdk@0.6.111
 - @ballerine/workflow-browser-sdk@0.6.111
 
 ## 0.3.109